Move JXMLComponent to its own file
